requirements.txt
# PyTorch (CPU-only for faster download, ~200MB vs 2GB CUDA version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
torchvision==0.16.2+cpu

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# AI/Computer Vision (will use torch CPU from above)
ultralytics>=8.3.0  # YOLOv11 support (C3k2 module)
opencv-python-headless==4.9.0.80
numpy==1.26.3
pillow==10.2.0

# Utilities
python-dotenv==1.0.1
